requirements.txt
flask==2.0.1
flask-cors==3.0.10
librosa==0.9.1
nemo-toolkit[all]==1.20.0